ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.0
python-docx==1.1.0
# spacy==3.7.0  # Optional — incompatible with Python 3.14 (pydantic v1 issue)
openai==1.50.0
scikit-learn==1.5.0
numpy==1.26.0
pydantic==2.9.0
requests==2.32.0
